feat: moved the status count to zustand store
